+++ Tests/Unit/Domain/Model/Event/EventDateTest.php
@@ -78,6 +78,14 @@ final class EventDateTest extends UnitTestCase
     /**
      * @test
      */
+    public function isRegistrationPossibleForThisClassReturnsTrue(): void
+    {
+        self::assertTrue($this->subject->isRegistrationPossibleForThisClass());
+    }
+
+    /**
+     * @test
+     */
     public function getInternalTitleInitiallyReturnsEmptyString(): void
     {
         self::assertSame('', $this->subject->getInternalTitle());